requirements.txt
fastapi
gunicorn
pydantic
PyYAML
requests
starlette
uvicorn
uvloop
aiohttp
async-timeout
